move on-chain data query to a separate hook
