+++ src/hooks/useConnect.ts
@@ -1,93 +1,22 @@
 import { useDisclosure } from '@chakra-ui/react'
-import { ApiPromise } from '@polkadot/api'
 import { web3Accounts, web3Enable } from '@polkadot/extension-dapp'
 import { encodeAddress } from '@polkadot/keyring'
-import { WsProvider } from '@polkadot/rpc-provider'
 import { useCallback, useEffect } from 'react'
-import { SUBSPACE_EXTENSION_ID, initialExtensionValues } from '../constants'
+import { SUBSPACE_ACCOUNT_FORMAT, SUBSPACE_EXTENSION_ID, initialExtensionValues } from '../constants'
 import { useExtension } from '../states/extension'
-import {
-  AccountDetails,
-  DomainRegistry,
-  DomainStakingSummary,
-  OperatorIdOwner,
-  Operators,
-  PendingStakingOperationCount
-} from '../types'
+import { AccountDetails } from '../types'
 
 export const useConnect = () => {
   const extension = useExtension((state) => state.extension)
   const api = useExtension((state) => state.api)
-  const setApi = useExtension((state) => state.setApi)
   const setExtension = useExtension((state) => state.setExtension)
   const setSubspaceAccount = useExtension((state) => state.setSubspaceAccount)
   const setInjectedExtension = useExtension((state) => state.setInjectedExtension)
   const setAccountDetails = useExtension((state) => state.setAccountDetails)
-  const setStakingConstants = useExtension((state) => state.setStakingConstants)
   const { isOpen: isConnectOpen, onOpen: onConnectOpen, onClose: onConnectClose } = useDisclosure()
 
   const handleConnect = useCallback(async () => {
     setExtension({ ...initialExtensionValues, loading: true })
-
-    try {
-      if (!process.env.NEXT_PUBLIC_PROVIDER_URL) throw new Error('NEXT_PUBLIC_PROVIDER_URL not set')
-
-      const wsProvider = new WsProvider(process.env.NEXT_PUBLIC_PROVIDER_URL)
-      const _api = await ApiPromise.create({ provider: wsProvider })
-      if (_api) {
-        console.log('Connection Success')
-        setApi(_api)
-        const { maxNominators, minOperatorStake, stakeEpochDuration, stakeWithdrawalLockingPeriod } =
-          _api.consts.domains
-        console.log('stakimg', _api.consts.domains)
-
-        const domainRegistry = await _api.query.domains.domainRegistry.entries()
-        console.log(
-          'domainRegistry',
-          domainRegistry.map((domain) => domain[1].toJSON() as DomainRegistry)
-        )
-
-        const domainStakingSummary = await _api.query.domains.domainStakingSummary.entries()
-        console.log(
-          'domainStakingSummary',
-          domainStakingSummary.map((domain) => domain[1].toJSON() as DomainStakingSummary)
-        )
-
-        const operatorIdOwner = await _api.query.domains.operatorIdOwner.entries()
-        console.log(
-          'operatorIdOwner',
-          operatorIdOwner.map((operator) => operator[1].toJSON() as OperatorIdOwner)
-        )
-
-        const operators = await _api.query.domains.operators.entries()
-        console.log(
-          'operators',
-          operators.map((operator) => operator[1].toJSON() as Operators)
-        )
-
-        const pendingStakingOperationCount = await _api.query.domains.pendingStakingOperationCount.entries()
-        console.log(
-          'pendingStakingOperationCount',
-          pendingStakingOperationCount.map((operator) => operator[1].toJSON() as PendingStakingOperationCount)
-        )
-
-        setStakingConstants({
-          maxNominators: Number(maxNominators.toString()),
-          minOperatorStake: BigInt(minOperatorStake.toString()),
-          stakeEpochDuration: Number(stakeEpochDuration.toString()),
-          stakeWithdrawalLockingPeriod: Number(stakeWithdrawalLockingPeriod.toString()),
-          domainRegistry: domainRegistry.map((domain) => domain[1].toJSON() as DomainRegistry),
-          domainStakingSummary: domainStakingSummary.map((domain) => domain[1].toJSON() as DomainStakingSummary),
-          operatorIdOwner: operatorIdOwner.map((operator) => operator[1].toJSON() as OperatorIdOwner),
-          operators: operators.map((operator) => operator[1].toJSON() as Operators),
-          pendingStakingOperationCount: pendingStakingOperationCount.map(
-            (operator) => operator[1].toJSON() as PendingStakingOperationCount
-          )
-        })
-      }
-    } catch (error) {
-      console.log(error)
-    }
 
     web3Enable(SUBSPACE_EXTENSION_ID)
       .then((injectedExtensions) => {
@@ -108,6 +37,7 @@ export const useConnect = () => {
             defaultAccount: accounts[0]
           }
         })
+        setSubspaceAccount(encodeAddress(accounts[0].address, SUBSPACE_ACCOUNT_FORMAT))
 
         console.log('accounts', accounts)
       })
@@ -115,7 +45,7 @@ export const useConnect = () => {
         console.error('Error with connect', error)
         setExtension({ error, loading: false, data: undefined })
       })
-  }, [setApi, setExtension, setInjectedExtension, setStakingConstants])
+  }, [setExtension, setInjectedExtension, setSubspaceAccount])
 
   const handleSelectFirstWalletFromExtension = useCallback(
     async (source: string) => {
@@ -130,7 +60,7 @@ export const useConnect = () => {
             defaultAccount: mainAccount
           }
         })
-        setSubspaceAccount(encodeAddress(mainAccount.address, 2258))
+        setSubspaceAccount(encodeAddress(mainAccount.address, SUBSPACE_ACCOUNT_FORMAT))
       }
       onConnectClose()
     },
@@ -149,7 +79,7 @@ export const useConnect = () => {
             defaultAccount: mainAccount
           }
         })
-        setSubspaceAccount(encodeAddress(mainAccount.address, 2258))
+        setSubspaceAccount(encodeAddress(mainAccount.address, SUBSPACE_ACCOUNT_FORMAT))
       }
     },
     [extension, setExtension, setSubspaceAccount]
